% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression2.R
\docType{data}
\name{depression2}
\alias{depression2}
\title{Depression Data 2 for Repeated Measures ANOVA (Long Format)}
\format{A data frame with 96 rows and 4 columns.}
\usage{
data("depression2")
}
\description{
The data correspond to an experiment in which a treatment for depression is
 studied. Two groups of patients (1: control / 2: treatment) have been
 followed at five different times (0: pre-test, 1: one month post-test, 3: 3
 months follow-up and 6: 6 months follow-up). The dependant variable is a
 depression score.

 Repeated measures ANOVA can be performed in order to determine the effect
 of the treatment and the effect of time on the depression score.
}
\examples{
data(depression2)
head(as.data.frame(depression2))
}
