% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfestime.R
\docType{data}
\name{selfestime}
\alias{selfestime}
\title{Self Estime Score Data for One-way Repeated Measures ANOVA}
\format{A data frame with 10 rows and 4 columns.}
\usage{
data("selfestime")
}
\description{
The dataset contains 10 individuals' self estime score on three time points
 during a specific diet to determine whether their self estime improved.

 One-way repeated measures ANOVA can be performed in order to determine the
 effect of time on the self estime score.
}
\examples{
data(selfestime)
head(selfestime)
}
