% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AirPassengersDf.R
\docType{data}
\name{AirPassengersDf}
\alias{AirPassengersDf}
\title{Air Passengers}
\format{
A data frame with 144 rows and 2 columns (Month and Passengers).
}
\usage{
data("AirPassengersDf")
}
\description{
This dataset contains the number of monthly air passengers from
1949 to 1960. It is taken from the R datasets package and formatted into long
data frame format.
}
\examples{
data("AirPassengersDf")
head(AirPassengersDf)
}
