% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anxiety.R
\docType{data}
\name{anxiety}
\alias{anxiety}
\title{Anxiety Data for Two-Way Mixed ANOVA}
\format{A data frame with 45 rows and 5 columns.}
\usage{
data("anxiety")
}
\description{
The data provide the anxiety score, measured at three time
 points, of three groups of individuals practicing physical exercises at
 different levels (grp1: basal, grp2: moderate and grp3: high)

 Two-way mixed ANOVA can be used to evaluate if there is interaction between
 group and time in explaining the anxiety score.
}
\examples{
data(anxiety)
head(as.data.frame(anxiety))
}
