% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression Data for Two Way Mixed ANOVA}
\format{
A data frame with 24 rows and 6 columns.
}
\usage{
data("depression")
}
\description{
The data correspond to an experiment in which a treatment for depression is
 studied. Two groups of patients - control ("ctr") and treatment
 ("treated") - have been followed at four different time points: t0
 (pre-test) and three post-test follow-up measurements (t1, t2, t3). The
 dependent variable is a depression score.

 Repeated measures ANOVA can be performed in order to determine the effect
 of the treatment and the effect of time on the depression score.
}
\examples{
data(depression)
head(as.data.frame(depression))
}
