% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/depression.R
\docType{data}
\name{depression}
\alias{depression}
\title{Depression Data for Two Way Mixed ANOVA}
\format{A data frame with 24 rows and 5 columns.}
\usage{
data("depression")
}
\description{
The data correspond to an experiment in which a treatment for depression is
 studied. Two groups of patients (1: control / 2: treatment) have been
 followed at five different times (0: pre-test, 1: one month post-test, 3: 3
 months follow-up and 6: 6 months follow-up). The dependent variable is a
 depression score.

 Repeated measures ANOVA can be performed in order to determine the effect
 of the treatment and the effect of time on the depression score.
}
\examples{
data(depression)
head(as.data.frame(depression))
}
