% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genderweight.R
\docType{data}
\name{genderweight}
\alias{genderweight}
\title{Weight Data By Gender for Two-Samples Mean Test}
\format{A data frame with 40 rows and 3 columns}
\usage{
data("genderweight")
}
\description{
Contains the weights by sex (M for male; F for
 female). The question is whether the average women’s weight differs from the
 average men’s weight?

 A two-samples independent t-test can be performed to answer to this question.
}
\examples{
data(genderweight)
head(as.data.frame(genderweight))
}
