% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/housetasks.raw.R
\docType{data}
\name{housetasks.raw}
\alias{housetasks.raw}
\title{Housetasks}
\format{A data frame with 1744 rows and 2 columns (tasks and status).}
\usage{
data("housetasks.raw")
}
\description{
A data frame containing the frequency of execution of 13 house tasks in the couple.
}
\examples{
data(housetasks.raw)
table(housetasks.raw)

}
