% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jobsatisfaction.R
\docType{data}
\name{jobsatisfaction}
\alias{jobsatisfaction}
\title{Job Satisfaction Data for Two-Way ANOVA}
\format{
A data frame with 58 rows and 4 columns.
}
\usage{
data("jobsatisfaction")
}
\description{
Contains the job satisfaction score organized by gender and education level.
}
\examples{
data(jobsatisfaction)
head(as.data.frame(jobsatisfaction))
}
