% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.R
\docType{data}
\name{mice}
\alias{mice}
\title{Mice Weight Data for One Sample Mean Test}
\format{A data frame with 10 rows and 2 columns}
\usage{
data("mice")
}
\description{
Contains the weight of 10 mice. The question is whether the
 average weight of the mice differs from 25g.

A one sample t-test can be performed to answer to this question.
}
\examples{
data(mice)
head(as.data.frame(mice))
}
