% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice2.R
\docType{data}
\name{mice2}
\alias{mice2}
\title{Mice Weight Data for Paired-Samples Mean Test}
\format{
A data frame with 10 rows and 3 columns.
 \describe{
   \item{id}{mouse identifier (1 to 10).}
   \item{before}{the mouse weight before the treatment.}
   \item{after}{the mouse weight after the treatment.}
 }
}
\source{
A simulated dataset for teaching the paired-samples t-test.
}
\usage{
data("mice2")
}
\description{
contains the weight of 10 mice before and after the treatment.

A paired-samples t-test can be performed to answer to this question.

 It is the data set used in the Datanovia tutorial \dQuote{T-Test in R:
 One-Sample, Independent (Student & Welch) & Paired}
 (\url{https://www.datanovia.com/learn/biostatistics/two-groups/t-test-in-r}).
}
\examples{
data(mice2)
head(mice2)

# Paired-samples t-test: did the treatment change the weight?
t.test(mice2$before, mice2$after, paired = TRUE)
}
\seealso{
\code{\link{mice}}, the one-sample companion data set.
 Datanovia tutorial:
 \href{https://www.datanovia.com/learn/biostatistics/two-groups/t-test-in-r}{T-Test in R: One-Sample, Independent (Student & Welch) & Paired}.
}
