% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice2.R
\docType{data}
\name{mice2}
\alias{mice2}
\title{Mice Weight Data for Paired-Samples Mean Test}
\format{A data frame with 10 rows and 3 columns}
\usage{
data("mice2")
}
\description{
contains the weight of 10 mice before and after the treatment.

A paired-samples t-test can be performed to answer to this question.
}
\examples{
data(mice2)
head(as.data.frame(mice2))
}
