% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\docType{data}
\name{performance}
\alias{performance}
\title{Performance Data for Three-Way Mixed ANOVA}
\format{
A data frame with 60 rows and 5 columns.
}
\usage{
data("performance")
}
\description{
Contains the performance score measures of participants at two
 time points. The aim of this study is to evaluate the effect of gender and
 stress on performance score. The three-way mixed ANOVA test can be used to
 investigate this question.

 The data include two between-subjects factors (gender and stress) and one
 within-subject factor (time, repeated measures).
}
\examples{
data(performance)
head(as.data.frame(performance))
}
