% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/properties.R
\docType{data}
\name{properties}
\alias{properties}
\title{Properties Data for Chi-square Test of Independence}
\format{A data frame with 333 rows and 2 columns.}
\usage{
data("properties")
}
\description{
Contains the type of properties and the buyer types. Buyer
 categories are: "single male", "single female", "married couple" and
 "family".

 The type of property these buyers purchased were sorted into four
 categories: "flat", "bungalow" (i.e., a one-storey home), "detached house"
 and "terrace" (i.e., a block of adjoining houses).

 Chi-square test of independence can be used to assess the association between
 the type of buyer who purchases a property and the type of property that is
 purchased.
}
\examples{
data("properties")
head(as.data.frame(properties))
}
