% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress.R
\docType{data}
\name{stress}
\alias{stress}
\title{Stress Data for Two-Way ANCOVA}
\format{A data frame with 60 rows and 5 columns.}
\usage{
data("stress")
}
\description{
Researchers want to evaluate the effect of a new "treatment" and
 "exercise" on the stress score reduction after adjusting for "age".

 Two-way ANCOVA can be performed in order to determine
 whether there is interaction between exercise and treatment on the stress
 score.
}
\examples{
data(stress)
head(as.data.frame(stress))
}
