% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taskachievment.R
\docType{data}
\name{taskachievment}
\alias{taskachievment}
\title{Task Achievment Data for Cochran's Q Test}
\format{A data frame with 73 rows and 4 columns.}
\usage{
data("taskachievment")
}
\description{
Repeated measures nominal designs where 73 subjects are asked to
 perform 3 tasks. The outcome of each task is a dichotomous value, success or
 failure. Each row correspond to a participant (called "block" in the
 jargon). This a demo dataset for Cochran's Q test.
}
\examples{
data(taskachievment)
head(taskachievment)
}
