% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/titanic.raw.R
\docType{data}
\name{titanic.raw}
\alias{titanic.raw}
\title{Survival of Passengers on the Titanic}
\format{A data frame with 2201 rows and 4 columns.}
\usage{
data("titanic.raw")
}
\description{
Survival of passengers on the Titanic. This data set provides
 information on the fate of passengers on the fatal maiden voyage of the
 ocean liner ‘Titanic’. Columns are economic status (Class), Sex,
 Age and Survived.
}
\examples{
data(titanic.raw)
data("titanic.raw")
with(titanic.raw, table(Class, Survived))

}
